add common meta form components, refactor
