--- website/frontend/src/components/blocks/StatusPanel/StatusPanel.scss
@@ -1,42 +1,0 @@
-@import '../../../assets/scss/variables';
-
-.status-panel-block {
-  align-items: center;
-  background-color: $textColor;
-  border-radius: $borderRadiusBase;
-  bottom: 0;
-  display: flex;
-  gap: 16px;
-  margin-bottom: 28px;
-  left: 50%;
-  transform: translateX(-50%);
-  padding: 13px 16px;
-  position: fixed;
-  z-index: 10;
-  &.error {
-    background-color: $errorColor;
-    .status-panel-block__msg {
-      color: $textColor !important;
-    }
-  }
-}
-
-.status-panel-block__msg {
-  color: $bkgInputColor;
-}
-
-.status-panel-block__close-btn {
-  background: url('../../../assets/images/close-success.svg');
-  &.error {
-    background: url('../../../assets/images/close.svg');
-  }
-  height: 14px;
-  width: 14px;
-}
-
-@media (max-width: 768px) {
-  .status-panel-block {
-    width: 92%;
-    justify-content: space-between;
-  }
-}
