+++ website/frontend/src/components/blocks/UploadMetaForm/index.ts
@@ -1,0 +1,1 @@
+export * from "./UploadMetaForm";
