+++ website/frontend/src/components/common/MetaFile/MetaFile.module.scss
