+++ website/frontend/src/components/common/MetaFile/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetaFile";
