+++ website/frontend/src/components/common/MetaSwitch/index.ts
@@ -1,0 +1,1 @@
+export * from "./MetaSwitch";
