--- website/frontend/src/components/pages/Meta/Meta.module.scss
@@ -1,10 +1,0 @@
-@use 'assets/scss/variables' as *;
-@use 'assets/scss/breakpoints.scss' as *;
-
-.metaFormWrapper {
-   background-color: $bkgActiveButtonColor;
-   border-radius: 2px;
-   border: 1px dashed #0e0e0e;
-   margin-bottom: 40px;
-   padding: 30px 25px;
-}
